+++ include/core/commands.h
@@ -251,6 +251,7 @@ void handle_comm_send(int argc, char **argv);
 void handle_comm_status(int argc, char **argv);
 void handle_comm_disconnect(int argc, char **argv);
 void handle_comm_setpins(int argc, char **argv);
+void handle_glbench_cmd(int argc, char **argv);
 void cmd_comm_register_callback(void);
 
 // GhostLink peer-flashing commands (see managers/peer_ota_manager.c)