+++ main/managers/settings_manager.c
@@ -1295,6 +1295,7 @@ void settings_persist_setting(SettingsType setting) {
 #endif
         case SETTING_RUN_SETUP_WIZARD:
         case SETTING_I2C_SCAN:
+        case SETTING_GL_BENCH:
         case SETTING_WIGLE_TEST_API:
         case SETTING_WIGLE_HELP:
         case SETTING_WIGLE_MANUAL_UPLOAD: